Import camp-engine.
Initial development was done in another repo to try to ensure
low coupling with the rest of the project so that it might one
day be used in other contexts (e.g. on the frontend). But for now,
having it in another repo just complicates our development and
build process, so move it in.