Allow SphereChoice to select a sphere multiple times.
A new ChoiceDef flag, `multi` signals that a choice can
potentially accept a selection more than once. This is stored
in an annoying way in the model. For example, if 'arcane' is
selected 3 times, the string 'arcane:3' will be recorded in the
choice model.

To go with this, display the character's spell slots on the
character details page.

Fixes #40
Fixes #77