Implement Opus.
This allows `artisan.specialization+crafter` etc to be used as
a requirement. The value of these expressions will be the number
of powers held with that tag, but only if that tag is the current
specialization.

Also adds a `grant_if` field for Tempest feature definitions
of the form:

```yaml
grant_if:
    feature-a: requirements-for-a
    feature-b: requirements-for-b
```