Enable buying Lores in certain edge cases.
If a Lore skill is individually discounted, such as from a
Culture, and the character has insufficient CP to buy Lores
in general but sufficient to buy the discounted one, the
discounted Lores will be shown by themselves.

(This applies to any option skill where individual discounts
are granted, but that's rare outside of Lore).