Initial event registration.
From the event page, you will see:

1. A note about your registration status.
2. If the registration window is open, a button to register or edit registration.

Whether a registration is new or being edited, the form is a little barebones right now.
You can choose to register as an NPC or select a character. The character must be:

a) One of _your_ characters.
b) Part of this campaign.

If you don't have a character that qualifies, the registration form will generate
one as a part of registration and, if you're PCing, punt you over there to do the work.

There's also a generic details textarea that is initially populated with whatever the event's details template looks like.

For #18.