Don't allow registration or updates after the registration window closes.
If you view your registration outside of the window, the form
elements will be disabled and an alert is displayed related to this.

If you force the form to submit anyway, a failure message is displayed.

Minor update: After successful register/edit, redirects back to the event. Except in the "new PC needs a character" case, where you'll be punted to the character editor.

For #18.